Completion of most recent challenge
I passes all the grunt tests but I have a very minimal
understanding of what's going on here in this program.
If you could help expand on important things going on here that
make this to work as well as some recommended readings to understand
this challenge more that would be greatly appreciated.